Реализован регистрация клиента.

+++ src/main/resources/db/migration/V1.0.1__create_table_login.sql
@@ -5,7 +5,7 @@ CREATE TABLE Login(
 	username varchar(256),
 	password varchar(256),
 	roles_id integer,
-	customer_id integer,
+	customer_id bigserial,
 	seller_company_id bigserial,
-	shipper_id integer
+	shipper_id bigserial
 );
